requirements.txt
scrapegraphai==1.11.3
fastapi==0.111.1
python-dotenv==1.0.1
nest-asyncio==1.6.0